Added explicit border conditions (absorbing/reflective/mainland)
This is a major overhaul of the dispersal code. The original code was
quite a mess, hard to understand and with at least two
bugs. Critically, it included assumptions about the model map and
study intentions that weren't met anymore, leading to unexpected
behaviour (periodic rather than absorbing borders).

This commit and the one before it address these issues. They
hopefully make the code a bit more readable (at least it's better
commented now) and fixed aforesaid bugs. They also remove the implicit
assumption of island/mainland scenarios and make absorbing borders the
*de facto* default. The former model behaviour is still accessible by
 setting border conditions to "mainland". For backward compatibility
 with old config files,  using `static = true` will automatically
 select mainland border conditions.